move db to server/db
