+++ .env.example
@@ -11,5 +11,5 @@
 
 NODE_ENV="development"
 # Turso
-TURSO_DATABASE_URL="file:./src/server/local.db"
+TURSO_DATABASE_URL="file:./src/server/db/local.db"
 TURSO_AUTH_TOKEN=""
